Fix #1: stopAnimation now emits signal & cleans up
This change makes it so that we internally track the timer state as a
new Animation property called timer_running. Internal tracking is used
instead of observing a property of self.timer, because the "timer" in
this case is actually just an ID.

The type check against self.timer has also been removed, because
self.timer will always either be of type nil or string, since it is
exclusively populated by the return value of glib.timeout_add.

These two changes make it possible to properly emit the
"anim::animation_stoped" signal and clean up the glib source when
animations are forcibly stopped via stopAnimation